Add interpreter class
