+++ cinder/visitors/__init__.py
@@ -1,3 +1,3 @@
 from cinder.visitors.compiler import ASTCompiler
 from cinder.visitors.verify import ASTVerifier
-from cinder.visitors.visitor import Visitor
+from cinder.visitors.visitor import Interpreter, Visitor
